Issue #70
When the "exclude_from_localizations" is used the list of paths was being passed to the Jekyll "exclude" list while generating the translated subsites, for this reason the files were not being copied to the subsites as intended, but as a side effect, the "site.static_files" also got empty and it's data that should be available to the subsites was inexistent, even though, the subsites are able to access the static files residing on the main site path.

This problem was solved by instead of adding the paths from "exclude_from_localizations" into Jekyll's "exclude", the plugin now uses a Jekyll Hook that runs right before Jekyll writes the site files. This Hook removes any file from the list of static files that will be written, if their relative path matches a path expansion from "exclude_from_localizations".