Apply refactoring and simplify code
A bug in 'serve' mode of Jekyll was fixed. The bug was in empty strings
once site rebuilt because of an attempt to translate already translated
strings once again.
Class Localizer has been removed due to redundancy and has been replaced
with function 'translate_props'.
New class TranslatedString was instroduced to store a key within
translated value and distinguish key and translated value.