Bug fix for page_url
https://github.com/jekyll/jekyll/pull/7117 describes an issue I noted with the `jekyll-multiple-languages-plugin` w/ regards to the `page_url` tag. It generated deprecation warnings.

That PR was rejected because the Jekyll maintainers felt it was the wrong way to change Jekyll core functionality because of details in how this plugin works. So, here I my suggested workaround for now. Because the regex is set up in the constructor, I had to include the whole constructor code from `Jekyll::Tags::PostComparer` inline here and just change the parts we want to override. Not very pretty but I don't see any obvious, easier way to get it sorted out for now.